Add GPU index support
This commit allows users to specify the GPU index when running the
execute.py script. The GPU index can be passed as a command-line argument
after the address.

The perform function has been updated to automatically detect the available
CUDA devices and allow the user to select one by providing the index as a
command-line argument. If no argument is provided, it will use the default
GPU (index 0). If no CUDA devices are available, it will use the CPU.

The Makefile has also been updated to include an optional gpu variable,
which can be passed to the execute.py script when running the make run
target.