Clarified print statement to user during error state
Problem:
The user was priorly getting an error message that said it would retry in 10 seconds... however, it is possible it takes 10 *MINUTES*. 

Why:
Because of how the loop in the code works, if the network connection is faulty, the entire retry can take up to 10 minutes. The 10 minutes comes from the "timeout" argument being set to 600 seconds in the "response = requests.post()" call.

User effect:
This situation results in the user 1) being confused and 2) not adequately understanding the unfolding erroneous state. This can leave users frustrated, and unsure of how to solve it.

Fix:
While simple, this fix 1) corrects the timing expectations (from 10 seconds to 10 mins) and 2) gives a hint as to what the program is trying to do and what it requires.